+++ src/main/java/br/com/cepp/maps/financas/service/AtivoValorService.java
@@ -6,6 +6,7 @@ import br.com.cepp.maps.financas.repository.AtivoValorRepository;
 import br.com.cepp.maps.financas.resource.dto.AtivoValorRequestDTO;
 import br.com.cepp.maps.financas.resource.handler.AtivoValorJaExisteException;
 import br.com.cepp.maps.financas.resource.handler.AtivoValorNaoEncontradoException;
+import br.com.cepp.maps.financas.resource.handler.AtivoValorUtilizadoException;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +17,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
+import java.util.ArrayList;
 
 @Log4j2
 @Service
@@ -37,7 +39,8 @@ public class AtivoValorService {
         }
 
         final Ativo ativo = this.ativoService.buscarPorCodigo(requestDTO.getCodigoAtivo());
-        final AtivoValor ativoValor = new AtivoValor(null, requestDTO.getData(), requestDTO.getValor(), ativo);
+        final AtivoValor ativoValor = new AtivoValor(null, requestDTO.getData(), requestDTO.getValor(), ativo,
+                new ArrayList<>());
 
         return this.repository.save(ativoValor);
     }
@@ -46,12 +49,22 @@ public class AtivoValorService {
     public void remover(@NotEmpty(message = "Campo 'codigoAtivo' é obrigatório") final String codigoAtivo,
                         @NotNull(message = "Campo 'data' é obrigatório") final LocalDate data) {
         final AtivoValor ativoValor = this.buscarPorAtivoEData(codigoAtivo, data);
+
+        if(!this.repository.existsByAtivo_CodigoAndDataAndMovimentosIsEmpty(codigoAtivo, data)) {
+            throw new AtivoValorUtilizadoException(codigoAtivo, data);
+        }
+
         this.repository.delete(ativoValor);
     }
 
     public AtivoValor buscarPorAtivoEData(@NotEmpty(message = "Campo 'codigoAtivo' é obrigatório") final String codigoAtivo,
-                                                    @NotNull(message = "Campo 'data' é obrigatório") final LocalDate data) {
+                                          @NotNull(message = "Campo 'data' é obrigatório") final LocalDate data) {
         return this.repository.findByAtivo_CodigoAndData(codigoAtivo, data)
                 .orElseThrow(() -> new AtivoValorNaoEncontradoException(codigoAtivo, data));
+    }
+
+    public boolean existsPorAtivoEData(@NotEmpty(message = "Campo 'codigoAtivo' é obrigatório") final String codigoAtivo,
+                                       @NotNull(message = "Campo 'data' é obrigatório") final LocalDate data) {
+        return this.repository.existsAtivoValorByAtivo_CodigoAndData(codigoAtivo, data);
     }
 }
